Moved first few problems to separate files
