+++ diffeqzoo/ivps/__init__.py
@@ -1,0 +1,6 @@
+"""Initial value problems."""
+
+
+from diffeqzoo.ivps._ivps import *
+from diffeqzoo.ivps._misc import logistic
+from diffeqzoo.ivps._oscillators import fitzhugh_nagumo, lotka_volterra
